[highlighting] Include third-party syntaxes as submodules.
All four of these projects (better-cpp-syntax, vscode-rust, TypeScript-TmLanguage, and MagicPython) include a copy of the MIT license in the root of their repository, and none place further limitations on derivative works related to their syntaxes in addition to those specified in the license. better-cpp-syntax borrows from another project that also uses the language of the MIT license.