Adjustments to header in small screens.
This fixes a bug where the icon does not change to the small version
when the screen starts out large and resizes to become small. It will
not work properly on Internet Explorer, but I do not think we need to
impress anyone who thinks it is OK to use IE.

It also removes overflow-x: scroll and uses flex-wrap instead.