Fix bottom nav for mobile.
This addresses the following three issues, all of which result from the
excessive area given to the bottom nav in the current version of the
website:
- The area devoted to the bottom nav is excessive. After this commit,
  two rows will not be used. Instead, the nav will overflow, and it will
  be possible to scroll to the right to view the other menu options.
  While scrolling to the right, a scroll bar will appear.
- The search bar appears but is partially obscured when not in use.
  After this commit, it will be completely obscured when not in use.
- The hamburger (three lines) icon used in the handbook section is
  obscured by the bottom nav. With this commit, it will no longer be
  obscured.

This commit also changes the width of the "search" svg from 16px to 20px
and slightly adjusts its placement.

It could be argued that the search box should be a few pixels smaller.
That can probably be done for the mobile view if it is important to us.